apps/api/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic and settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP client for OAuth
httpx==0.26.0

# Logging and monitoring
python-json-logger==2.0.7
structlog==24.1.0

# Rate limiting
slowapi==0.1.9

# CORS and middleware
python-cors==1.0.0

# WebSocket
python-socketio==5.10.0
python-engineio==4.8.0
aioredis==2.0.1

# Utilities
python-dotenv==1.0.0
